LR/requirements.txt
astropy==7.1.0
PyYAML==6.0.2
